+++ NEWS.md
@@ -25,8 +25,10 @@ ParseHeaders:
   the existing field.
 + Added the copy subcommand which will copy annotations into new field names
   or merge the annotations of existing fields.
-+ Added a commandline check to ensure that the `-f` and `-k` arguments contain
-  the same number of fields for both the rename and copy subcommands.
++ Added the `--act` argument to the copy and rename subcommands allowing 
+  collapse following the copy or rename operation.
++ Added a commandline check to ensure that the `-f`, `-k` and `--act` arguments 
+  contain the same number of fields for both the rename and copy subcommands.
 
 
 Version 0.4.7:  June 5, 2015